+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod aliases;
 pub mod app;
 pub mod cache;
 pub mod cli;